Fix `assert_called_with` with an array as expected argument
Fixes #4

There is a suggestion to change API of `assert_called_with` to
allow providing a different return value for each set of arguments.
It is a superb suggestion, but I currently hesitate to change the API in
that way. Let's not rush with changing the API.

```ruby
expectations = [
  {
    expected_args: ["Thanks for sharing this."],
    return_value: mocked_comment_1
  },
  {
    expected_args:  ["Thanks!"],
    return_value: mocked_comment_2
  }
]

assert_called_with(@post, :add_comment, expectations: expectations) do
  @post.add_comment("Thanks for sharing this.")
  @post.add_comment("Thanks!")
end
```

Thank you @thatguysimon for finding that bug, reporting the issue and
suggesting to imporve the API of `assert_called_with`, I appreciate that.